requirements.txt
kivy
kivymd
cvzone
mediapipe
opencv-python
psycopg2
bcrypt
google_auth_oauthlib
pygame
deepface
tensorflow
tf-keras
python-socketio
pycaw
eventlet
flask
flask-socketio
numpy
shutil
ctypes
comtypes
collections
smtplib
python-dotenv